Use the glib integer/pointer conversion macros for 64 bit cleanliness.
2003-11-13  Joe Shaw  <joe@ximian.com>

	* rcd-rpc-prefs.c (get_boolean_func, get_int_func,
	set_boolean_func, set_int_func): Use the glib integer/pointer
	conversion macros for 64 bit cleanliness.

	* rcd-unix-server.c (read_data): Fix "int" to "gsize" for sizes
	passed into g_io_channel_{read_write} so that things work nicely
	on 64 bit archs.

+++ src/rcd-unix-server.c
@@ -93,11 +93,11 @@ read_data(GIOChannel *iochannel,
     RCDUnixServerHandle *handle = user_data;
     GIOError err;
     char read_buf[4096];
-    int bytes_read = 0;
-    int total_read = 0;
+    gsize bytes_read = 0;
+    gsize total_read = 0;
     GByteArray *result;
-    int bytes_written;
-    int total_written = 0;
+    gsize bytes_written;
+    gsize total_written = 0;
 
 #ifdef HAVE_SO_PEERCRED
     if (!handle->cred_available)
